backend/requirements.txt
Django==6.0.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
psycopg[binary]==3.3.2
python-dotenv==1.2.1
uvicorn==0.40.0
pytest==8.3.4
pytest-django==4.9.0